chore(components): move icon files to Icons folder
