+++ __tests__/Navbar.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { Navbar } from '../components/layout/Navbar';
-import ArchiveIcon from '../components/ArchiveIcon';
-import UserIcon from '../components/UserIcon';
-import BackArrowIcon from '../components/BackArrowIcon';
+import ArchiveIcon from '../components/Icons/ArchiveIcon';
+import UserIcon from '../components/Icons/UserIcon';
+import BackArrowIcon from '../components/Icons/BackArrowIcon';
 
 describe('Navbar', () => {
   it('renders the title', () => {
